+++ NEWS.md
@@ -1,3 +1,8 @@
+* added `paired_test_continuous()` function to implement paired test
+  for continuous variable accordingly to the number of groups, in a 
+  suitable way to be used in the **Hmisc** `summary()` with
+  `method = "reverse"`.
+
 # depigner 0.1.0
 
 * some variable names updated